# models/Zuct_db/generated/CoursesModelGenerated.js
/**
 * 
 * 
  _____                      _              _ _ _     _   _     _        __ _ _      
 |  __ \                    | |            | (_) |   | | | |   (_)      / _(_) |     
 | |  | | ___    _ __   ___ | |_    ___  __| |_| |_  | |_| |__  _ ___  | |_ _| | ___ 
 | |  | |/ _ \  | '_ \ / _ \| __|  / _ \/ _` | | __| | __| '_ \| / __| |  _| | |/ _ \
 | |__| | (_) | | | | | (_) | |_  |  __/ (_| | | |_  | |_| | | | \__ \ | | | | |  __/
 |_____/ \___/  |_| |_|\___/ \__|  \___|\__,_|_|\__|  \__|_| |_|_|___/ |_| |_|_|\___|
 
 * DO NOT EDIT THIS FILE!! 
 * 
 *  TO CUSTOMIZE coursesModelGenerated.js PLEASE EDIT ../coursesModel.js
 * 
 *  -- THIS FILE WILL BE OVERWRITTEN ON THE NEXT SKAFFOLDER'S CODE GENERATION --
 * 
 */
// Database
import Database from "../../../classes/Database_Zuct_db";
import mongoose, { Schema } from "mongoose";

// Logger
import Logger from "../../../classes/Logger";

const generatedModel = {
  /**
   * Init  schema
   */
  init() {
    const db = Database.getConnection();

    /**
      * courses
      */
    const coursesSchema = new mongoose.Schema({
      course_code: {
        type: "String", 
        required: true
      },
      course_name: {
        type: "String", 
        required: true
      },
      // RELATIONS
      _contains: {
        type: Schema.ObjectId,
        ref: "programs"
      },
      _teaches: {
        type: Schema.ObjectId,
        ref: "lecturers"
      },
      
      
      // EXTERNAL RELATIONS
      /*
      _takes: [{
        type: Schema.ObjectId,
        ref: "students"
      }],
      */
    });

    generatedModel.setModel(db.connection.model("Courses", coursesSchema));

    return coursesSchema;
  },

  /**
   * Set Model
   */
  setModel: model => {
    generatedModel.model = model;
  },

  /**
   * Get model
   */
  getModel: () => {
    return generatedModel.model;
  },

  // Start queries
    

  // CRUD METHODS


  /**
  * coursesModel.create
  *   @description CRUD ACTION create
  *
  */
  async create(item) {
    const obj = new generatedModel.model(item);
    return await obj.save();
  },
  
  /**
  * coursesModel.delete
  *   @description CRUD ACTION delete
  *   @param ObjectId id Id
  *
  */
  async delete(id) {
    return await generatedModel.model.findByIdAndRemove(id);
  },
  
  /**
  * coursesModel.findBy_contains
  *   @description CRUD ACTION findBy_contains
  *   @param Objectid key Id of model to search for
  *
  */
  async findBy_contains(key) {
    return await generatedModel.model.find({ '_contains' : key});
  },
  
  /**
  * coursesModel.findBy_teaches
  *   @description CRUD ACTION findBy_teaches
  *   @param Objectid key Id of model to search for
  *
  */
  async findBy_teaches(key) {
    return await generatedModel.model.find({ '_teaches' : key});
  },
  
  /**
  * coursesModel.get
  *   @description CRUD ACTION get
  *   @param ObjectId id Id resource
  *
  */
  async get(id) {
    return await generatedModel.model.findOne({ _id : id });
  },
  
  /**
  * coursesModel.list
  *   @description CRUD ACTION list
  *
  */
  async list() {
    return await generatedModel.model.find();
  },
  
  /**
  * coursesModel.update
  *   @description CRUD ACTION update
  *   @param ObjectId id Id
  *
  */
  async update(item) { 
    return await generatedModel.model.findOneAndUpdate({ _id: item._id }, item, {'new': true});
  },
  


};

export default generatedModel;