# models/Zuct_db/generated/LecturersModelGenerated.js
/**
 * 
 * 
  _____                      _              _ _ _     _   _     _        __ _ _      
 |  __ \                    | |            | (_) |   | | | |   (_)      / _(_) |     
 | |  | | ___    _ __   ___ | |_    ___  __| |_| |_  | |_| |__  _ ___  | |_ _| | ___ 
 | |  | |/ _ \  | '_ \ / _ \| __|  / _ \/ _` | | __| | __| '_ \| / __| |  _| | |/ _ \
 | |__| | (_) | | | | | (_) | |_  |  __/ (_| | | |_  | |_| | | | \__ \ | | | | |  __/
 |_____/ \___/  |_| |_|\___/ \__|  \___|\__,_|_|\__|  \__|_| |_|_|___/ |_| |_|_|\___|
 
 * DO NOT EDIT THIS FILE!! 
 * 
 *  TO CUSTOMIZE lecturersModelGenerated.js PLEASE EDIT ../lecturersModel.js
 * 
 *  -- THIS FILE WILL BE OVERWRITTEN ON THE NEXT SKAFFOLDER'S CODE GENERATION --
 * 
 */
// Database
import Database from "../../../classes/Database_Zuct_db";
import mongoose, { Schema } from "mongoose";

// Logger
import Logger from "../../../classes/Logger";

const generatedModel = {
  /**
   * Init  schema
   */
  init() {
    const db = Database.getConnection();

    /**
      * lecturers
      */
    const lecturersSchema = new mongoose.Schema({
      first_name: {
        type: "String", 
        required: true
      },
      gender: {
        type: "String", 
        required: true
      },
      last_name: {
        type: "String", 
        required: true
      },
      middle_name: {
        type: "String"
      },
      phone: {
        type: "Number"
      },
      // RELATIONS
      _contain: {
        type: Schema.ObjectId,
        ref: "programs"
      },
      
      
      // EXTERNAL RELATIONS
      /*
      _taught: [{
        type: Schema.ObjectId,
        ref: "students"
      }],
      _teaches: {
        type: Schema.ObjectId,
        ref: "courses"
      },
      */
    });

    generatedModel.setModel(db.connection.model("Lecturers", lecturersSchema));

    return lecturersSchema;
  },

  /**
   * Set Model
   */
  setModel: model => {
    generatedModel.model = model;
  },

  /**
   * Get model
   */
  getModel: () => {
    return generatedModel.model;
  },

  // Start queries
    

  // CRUD METHODS


  /**
  * lecturersModel.create
  *   @description CRUD ACTION create
  *
  */
  async create(item) {
    const obj = new generatedModel.model(item);
    return await obj.save();
  },
  
  /**
  * lecturersModel.delete
  *   @description CRUD ACTION delete
  *   @param ObjectId id Id
  *
  */
  async delete(id) {
    return await generatedModel.model.findByIdAndRemove(id);
  },
  
  /**
  * lecturersModel.findBy_contain
  *   @description CRUD ACTION findBy_contain
  *   @param Objectid key Id of model to search for
  *
  */
  async findBy_contain(key) {
    return await generatedModel.model.find({ '_contain' : key});
  },
  
  /**
  * lecturersModel.get
  *   @description CRUD ACTION get
  *   @param ObjectId id Id resource
  *
  */
  async get(id) {
    return await generatedModel.model.findOne({ _id : id });
  },
  
  /**
  * lecturersModel.list
  *   @description CRUD ACTION list
  *
  */
  async list() {
    return await generatedModel.model.find();
  },
  
  /**
  * lecturersModel.update
  *   @description CRUD ACTION update
  *   @param ObjectId id Id
  *
  */
  async update(item) { 
    return await generatedModel.model.findOneAndUpdate({ _id: item._id }, item, {'new': true});
  },
  


};

export default generatedModel;